project ver6/project ver5/project/pep272-encryption-master/pep272-encryption-master/requirements.txt
pycryptodome
pytest>=3.0.0
pytest-runner>=2.0.0
setuptools>=27.3
wheel
sphinx
sphinx_rtd_theme
coverage